HstMcl3: refactoring non matching acceptance test with MSA metadata dependencies
Making MSA data optional in the VSP configuration optional by making it nullable
in the case of non matching https://trello.com/c/NHstMcl3/242-make-msa-metadata-optional
Acceptance criteria is when all non-matching test that contain MSA configuration can
be made to pass without requiring MSA configuration data. Steps involved requires

- Removing multiple classes for non matching acceptance test
- Ensuring there is only one non matching acceptance test class
- And therefore only one place to MSA configuration data to change
- make MSA metadata optional and nullable
