--- src/acceptance-test/java/uk/gov/ida/verifyserviceprovider/NonMatchingAuthnFailedResponseAcceptanceTest.java
@@ -1,73 +1,0 @@
-package uk.gov.ida.verifyserviceprovider;
-
-import com.google.common.collect.ImmutableMap;
-import common.uk.gov.ida.verifyserviceprovider.servers.MockMsaServer;
-import org.junit.Before;
-import org.junit.BeforeClass;
-import org.junit.ClassRule;
-import org.junit.Test;
-import uk.gov.ida.verifyserviceprovider.dto.NonMatchingScenario;
-import uk.gov.ida.verifyserviceprovider.dto.RequestResponseBody;
-import uk.gov.ida.verifyserviceprovider.dto.TestTranslatedNonMatchingResponseBody;
-import uk.gov.ida.verifyserviceprovider.rules.VerifyServiceProviderAppRule;
-import uk.gov.ida.verifyserviceprovider.services.ComplianceToolService;
-import uk.gov.ida.verifyserviceprovider.services.GenerateRequestService;
-
-import javax.ws.rs.client.Client;
-import javax.ws.rs.core.Response;
-import java.util.Map;
-
-import static javax.ws.rs.client.Entity.json;
-import static javax.ws.rs.core.Response.Status.OK;
-import static org.assertj.core.api.Assertions.assertThat;
-import static uk.gov.ida.verifyserviceprovider.builders.VerifyServiceProviderAppRuleBuilder.aVerifyServiceProviderAppRule;
-import static uk.gov.ida.verifyserviceprovider.dto.LevelOfAssurance.LEVEL_2;
-import static uk.gov.ida.verifyserviceprovider.services.ComplianceToolService.AUTHENTICATION_FAILED_WITH_NON_MATCH_SETTING_ID;
-
-public class NonMatchingAuthnFailedResponseAcceptanceTest {
-
-    @ClassRule
-    public static MockMsaServer msaServer = new MockMsaServer();
-
-    @ClassRule
-    public static VerifyServiceProviderAppRule application = aVerifyServiceProviderAppRule()
-            .withMockMsaServer(msaServer)
-            .build();
-
-    private static Client client;
-    private static ComplianceToolService complianceTool;
-    private static GenerateRequestService generateRequestService;
-
-    @BeforeClass
-    public static void setUpBeforeClass() {
-        client = application.client();
-        complianceTool = new ComplianceToolService(client);
-        generateRequestService = new GenerateRequestService(client);
-    }
-
-    @Before
-    public void setUp() {
-        complianceTool.initialiseWithDefaultsForV2();
-    }
-
-    @Test
-    public void shouldRespondWithSuccessWhenAuthnFailed() {
-        RequestResponseBody requestResponseBody = generateRequestService.generateAuthnRequest(application.getLocalPort());
-        Map<String, String> translateResponseRequestData = ImmutableMap.of(
-            "samlResponse", complianceTool.createResponseFor(requestResponseBody.getSamlRequest(), AUTHENTICATION_FAILED_WITH_NON_MATCH_SETTING_ID),
-            "requestId", requestResponseBody.getRequestId(),
-            "levelOfAssurance", LEVEL_2.name()
-        );
-
-        Response response = client
-            .target(String.format("http://localhost:%d/translate-non-matching-response", application.getLocalPort()))
-            .request()
-            .buildPost(json(translateResponseRequestData))
-            .invoke();
-
-        TestTranslatedNonMatchingResponseBody responseContent = response.readEntity(TestTranslatedNonMatchingResponseBody.class);
-
-        assertThat(response.getStatus()).isEqualTo(OK.getStatusCode());
-        assertThat(responseContent.getScenario()).isEqualTo(NonMatchingScenario.AUTHENTICATION_FAILED);
-    }
-}
