+++ src/acceptance-test/java/uk/gov/ida/verifyserviceprovider/rules/VerifyServiceProviderAppRule.java
@@ -7,7 +7,6 @@ import keystore.KeyStoreResource;
 import uk.gov.ida.saml.core.IdaSamlBootstrap;
 import uk.gov.ida.verifyserviceprovider.VerifyServiceProviderApplication;
 import uk.gov.ida.verifyserviceprovider.configuration.VerifyServiceProviderConfiguration;
-
 import static keystore.builders.KeyStoreResourceBuilder.aKeyStoreResource;
 import static uk.gov.ida.saml.core.test.TestCertificateStrings.METADATA_SIGNING_A_PUBLIC_CERT;
 import static uk.gov.ida.saml.core.test.TestCertificateStrings.TEST_RP_PRIVATE_ENCRYPTION_KEY;
@@ -24,6 +23,22 @@ public class VerifyServiceProviderAppRule extends DropwizardAppRule<VerifyServic
     }
 
 
+
+    public VerifyServiceProviderAppRule(String secondaryEncryptionKey, String serviceEntityIdOverride) {
+        super(
+            VerifyServiceProviderApplication.class,
+            "verify-service-provider.yml",
+            ConfigOverride.config("serviceEntityIds", serviceEntityIdOverride),
+            ConfigOverride.config("hashingEntityId", "some-hashing-entity-id"),
+            ConfigOverride.config("server.connector.port", String.valueOf(0)),
+            ConfigOverride.config("logging.loggers.uk\\.gov", "DEBUG"),
+            ConfigOverride.config("samlSigningKey", TEST_RP_PRIVATE_SIGNING_KEY),
+            ConfigOverride.config("verifyHubConfiguration.environment", "COMPLIANCE_TOOL"),
+            ConfigOverride.config("samlPrimaryEncryptionKey", TEST_RP_PRIVATE_ENCRYPTION_KEY),
+            ConfigOverride.config("samlSecondaryEncryptionKey", secondaryEncryptionKey),
+            ConfigOverride.config("msaMetadata.uri", "http://some-msa-uri")
+        );
+    }
     public VerifyServiceProviderAppRule(MockMsaServer msaServer, String secondaryEncryptionKey, String serviceEntityIdOverride) {
         super(
             VerifyServiceProviderApplication.class,
@@ -51,11 +66,15 @@ public class VerifyServiceProviderAppRule extends DropwizardAppRule<VerifyServic
         );
     }
 
+    public VerifyServiceProviderAppRule() {
+        this(TEST_RP_PRIVATE_ENCRYPTION_KEY, "http://verify-service-provider");
+    }
+
     public VerifyServiceProviderAppRule(MockMsaServer msaServer) {
-        this(msaServer, TEST_RP_PRIVATE_ENCRYPTION_KEY, "http://verify-service-provider");
+        this(msaServer,TEST_RP_PRIVATE_ENCRYPTION_KEY, "http://verify-service-provider");
     }
 
     public VerifyServiceProviderAppRule(MockMsaServer msaServer, String serviceEntityIdOverride) {
-        this(msaServer, TEST_RP_PRIVATE_ENCRYPTION_KEY, serviceEntityIdOverride);
+        this(msaServer,TEST_RP_PRIVATE_ENCRYPTION_KEY, serviceEntityIdOverride);
     }
 }
